alpr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
numpy==2.1.1
opencv-python-headless==4.10.0.84
easyocr==1.7.2
python-dotenv==1.0.1
